Add files via upload
This system is a Database Management System (DBMS) using SQL for analyzing Airbnb data. 
It involves of a structured database to store and manage information related to Airbnb listings, hosts, reviews, and other relevant data. The primary goal is to perform analysis and gain insights from this data for various purposes.